Add position tracking on dragged magnets
Position of magnets is now extracted and recorded in the console log
when drag n' dropped in preparation for updating the server when
they are moved around.
Also updated jQuery and jQueryUI to allow .on("dragstop"...) functions.
Renamed magnet.js to main.js for better convention.
Further code cleanup because readability.