Add magnet generation from test API and location updating to server
Magnets are now dynamically generated from a test RESTful API courtesy of http://rest.learncode.academy. In addition, when magnets are dragged they now update their data on the server accordingly. As the data on the server is cleared nightly, the test.json file is still useful for initializing a data set for use in testing.