Add on-demand polling and periodic refreshing of magnet data
Added the capability to periodically refresh magnet positions in case
they've been  moved by another user. It's ugly... but it's something.
Also added a button that allows users to poll the server for a list
of the words and  their locations on the server. Finally, changed the UI
in order to prevent excessive eye bleeding.